yeah, we're alive, baby

+++ Lab Task 2/2_Task2.cpp
@@ -4,8 +4,8 @@ using namespace std;
 int main(){
     int basicPay = 21500;
     int sale = 55697404;
-    double bonus = (sale*10.5)/100;
-    double netPay = basicPay + bonus;
+    float bonus = (sale*10.5)/100;
+    float netPay = basicPay + bonus;
     
     cout<<"The basic pay of the Salesman is "<< basicPay<< endl;
     cout<<"Bonus of the given month is "<< bonus<< endl;
